Fix in back_at_prompt check
When we get a line startign with '[' and the prompt is already in the
buffer the pattern matches and we get wrong return code and are not back
at the prompt.

We now check if we have a line in the buffer and if so, we are not at
the prompt. The promtp is not terminated by \n

Signed-off-by: Jonatan Schlag <jonatan.schlag@ipfire.org>